Tratando as exceções de deleção de User

+++ src/main/java/ProjetoNelio/service/UserService.java
@@ -2,8 +2,11 @@ package ProjetoNelio.service;
 
 import ProjetoNelio.entities.User;
 import ProjetoNelio.repository.UserRepository;
+import ProjetoNelio.service.exceptions.DatabaseException;
 import ProjetoNelio.service.exceptions.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -30,7 +33,15 @@ public class UserService {
     }
 
     public void delete(Long id){
-        userRepository.deleteById(id);
+        try {
+            if (!userRepository.existsById(id)) {
+                throw new ResourceNotFoundException(id);
+            }
+            userRepository.deleteById(id);
+        }
+        catch (DataIntegrityViolationException e){
+            throw new DatabaseException(e.getMessage());
+        }
     }
 
     public User update(Long id, User user){
